blog-backend-fastapi/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
python-jose
passlib[bcrypt]
pydantic
python-dotenv
